requirements.txt
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
PyWavelets==1.5.0
matplotlib==3.8.2
seaborn==0.13.2
python-dotenv==1.0.0
pytest==7.4.3
joblib==1.3.2
